+++ src/main/resources/application.properties
@@ -1,4 +1,4 @@
-server.port=8080
+server.port=3000
 
 spring.mvc.view.prefix=/WEB-INF/views/
 spring.mvc.view.suffix=.jsp
